requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
aenum==3.1.11
    # via gremlinpython
aiohttp==3.8.1
    # via gremlinpython
aiosignal==1.2.0
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==21.4.0
    # via aiohttp
bidict==0.22.0
    # via python-socketio
charset-normalizer==2.0.12
    # via aiohttp
click==8.1.3
    # via flask
dnspython==2.2.1
    # via eventlet
eventlet==0.33.1
    # via -r requirements.in
flask==2.1.2
    # via flask-socketio
flask-socketio==5.2.0
    # via -r requirements.in
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==1.1.2
    # via eventlet
gremlinpython==3.6.0
    # via -r requirements.in
idna==3.3
    # via yarl
importlib-metadata==4.11.4
    # via flask
isodate==0.6.1
    # via gremlinpython
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.1
    # via jinja2
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
nest-asyncio==1.5.5
    # via gremlinpython
python-engineio==4.3.2
    # via python-socketio
python-socketio==5.6.0
    # via flask-socketio
six==1.16.0
    # via
    #   eventlet
    #   isodate
websocket-client==1.3.2
    # via -r requirements.in
werkzeug==2.1.2
    # via flask
yarl==1.7.2
    # via aiohttp
zipp==3.8.0
    # via importlib-metadata
